+++ lib/src/platform_text_field.dart
@@ -86,7 +86,6 @@ abstract class _BaseData {
     this.readOnly,
     this.showCursor,
     this.textAlignVertical,
-    this.toolbarOptions,
     this.onTap,
     this.enableSuggestions,
     this.smartDashesType,
@@ -107,6 +106,9 @@ abstract class _BaseData {
     this.onTapOutside,
     this.spellCheckConfiguration,
     this.magnifierConfiguration,
+    this.contentInsertionConfiguration,
+    this.cursorOpacityAnimates,
+    this.undoController,
   });
 
   final Key? widgetKey;
@@ -155,7 +157,6 @@ abstract class _BaseData {
   final bool? showCursor;
   final TextAlignVertical? textAlignVertical;
   final GestureTapCallback? onTap;
-  final ToolbarOptions? toolbarOptions;
   final bool? enableSuggestions;
   final SmartDashesType? smartDashesType;
   final SmartQuotesType? smartQuotesType;
@@ -167,6 +168,9 @@ abstract class _BaseData {
   final TapRegionCallback? onTapOutside;
   final SpellCheckConfiguration? spellCheckConfiguration;
   final TextMagnifierConfiguration? magnifierConfiguration;
+  final ContentInsertionConfiguration? contentInsertionConfiguration;
+  final bool? cursorOpacityAnimates;
+  final UndoHistoryController? undoController;
 }
 
 class MaterialTextFieldData extends _BaseData {
@@ -206,11 +210,6 @@ class MaterialTextFieldData extends _BaseData {
     super.readOnly,
     super.showCursor,
     super.textAlignVertical,
-    @Deprecated(
-      'Use `contextMenuBuilder` instead. '
-      'This feature was deprecated after v3.3.0-0.5.pre.',
-    )
-        super.toolbarOptions,
     super.enableSuggestions,
     super.smartDashesType,
     super.smartQuotesType,
@@ -229,10 +228,14 @@ class MaterialTextFieldData extends _BaseData {
     super.onTapOutside,
     super.spellCheckConfiguration,
     super.magnifierConfiguration,
+    super.contentInsertionConfiguration,
+    super.cursorOpacityAnimates,
+    super.undoController,
     this.decoration,
     this.buildCounter,
     this.mouseCursor,
     this.onAppPrivateCommand,
+    this.canRequestFocus,
   });
 
   final InputDecoration? decoration;
@@ -240,6 +243,7 @@ class MaterialTextFieldData extends _BaseData {
   final InputCounterWidgetBuilder? buildCounter;
   final MouseCursor? mouseCursor;
   final AppPrivateCommandCallback? onAppPrivateCommand;
+  final bool? canRequestFocus;
 }
 
 class CupertinoTextFieldData extends _BaseData {
@@ -277,11 +281,6 @@ class CupertinoTextFieldData extends _BaseData {
     super.readOnly,
     super.showCursor,
     super.textAlignVertical,
-    @Deprecated(
-      'Use `contextMenuBuilder` instead. '
-      'This feature was deprecated after v3.3.0-0.5.pre.',
-    )
-        super.toolbarOptions,
     super.onTap,
     super.enableSuggestions,
     super.smartDashesType,
@@ -363,7 +362,6 @@ class PlatformTextField
   final bool? readOnly;
   final bool? showCursor;
   final TextAlignVertical? textAlignVertical;
-  final ToolbarOptions? toolbarOptions;
 
   final Color? cursorColor;
   final Brightness? keyboardAppearance;
@@ -390,6 +388,9 @@ class PlatformTextField
   final TapRegionCallback? onTapOutside;
   final SpellCheckConfiguration? spellCheckConfiguration;
   final TextMagnifierConfiguration? magnifierConfiguration;
+  final ContentInsertionConfiguration? contentInsertionConfiguration;
+  final bool? cursorOpacityAnimates;
+  final UndoHistoryController? undoController;
 
   static Widget _defaultMaterialContextMenuBuilder(
       BuildContext context, EditableTextState editableTextState) {
@@ -441,11 +442,6 @@ class PlatformTextField
     this.readOnly,
     this.showCursor,
     this.textAlignVertical,
-    @Deprecated(
-      'Use `contextMenuBuilder` instead. '
-      'This feature was deprecated after v3.3.0-0.5.pre.',
-    )
-        this.toolbarOptions,
     this.smartDashesType,
     this.smartQuotesType,
     this.selectionHeightStyle,
@@ -465,6 +461,9 @@ class PlatformTextField
     this.onTapOutside,
     this.spellCheckConfiguration,
     this.magnifierConfiguration,
+    this.contentInsertionConfiguration,
+    this.cursorOpacityAnimates,
+    this.undoController,
     this.material,
     this.cupertino,
   }) : keyboardType = keyboardType ??
@@ -527,7 +526,6 @@ class PlatformTextField
       readOnly: data?.readOnly ?? readOnly ?? false,
       showCursor: data?.showCursor ?? showCursor,
       textAlignVertical: data?.textAlignVertical ?? textAlignVertical,
-      toolbarOptions: data?.toolbarOptions ?? toolbarOptions,
       enableSuggestions: data?.enableSuggestions ?? true,
       smartQuotesType: data?.smartQuotesType ?? smartQuotesType,
       smartDashesType: data?.smartDashesType ?? smartDashesType,
@@ -558,6 +556,13 @@ class PlatformTextField
           data?.spellCheckConfiguration ?? spellCheckConfiguration,
       magnifierConfiguration:
           data?.magnifierConfiguration ?? magnifierConfiguration,
+      canRequestFocus: data?.canRequestFocus ?? true,
+      contentInsertionConfiguration:
+          data?.contentInsertionConfiguration ?? contentInsertionConfiguration,
+      cursorOpacityAnimates:
+          data?.cursorOpacityAnimates ?? cursorOpacityAnimates,
+      undoController: data?.undoController ?? undoController,
+      // toolbarOptions: Deprecated
     );
   }
 
@@ -625,7 +630,6 @@ class PlatformTextField
       readOnly: data?.readOnly ?? readOnly ?? false,
       showCursor: data?.showCursor ?? showCursor,
       textAlignVertical: data?.textAlignVertical ?? textAlignVertical,
-      toolbarOptions: data?.toolbarOptions ?? toolbarOptions,
       enableSuggestions: data?.enableSuggestions ?? true,
       smartQuotesType: data?.smartQuotesType ?? smartQuotesType,
       smartDashesType: data?.smartDashesType ?? smartDashesType,
@@ -655,6 +659,12 @@ class PlatformTextField
           data?.spellCheckConfiguration ?? spellCheckConfiguration,
       magnifierConfiguration:
           data?.magnifierConfiguration ?? magnifierConfiguration,
+      contentInsertionConfiguration:
+          data?.contentInsertionConfiguration ?? contentInsertionConfiguration,
+      cursorOpacityAnimates:
+          data?.cursorOpacityAnimates ?? cursorOpacityAnimates ?? true,
+      undoController: data?.undoController ?? undoController,
+      // toolbarOptions: Deprecated
     );
   }