+++ lib/src/platform_theme.dart
@@ -1,6 +1,5 @@
 import 'package:flutter/cupertino.dart' show CupertinoThemeData;
 import 'package:flutter/material.dart' show Theme, ThemeData, ThemeMode;
-import 'package:flutter/scheduler.dart';
 import 'package:flutter/services.dart';
 import 'package:flutter/widgets.dart';
 
@@ -97,7 +96,7 @@ class _PlatformThemeState extends State<PlatformTheme>
 
   bool get isDark {
     final platformBrightness =
-        SchedulerBinding.instance.window.platformBrightness;
+        View.of(context).platformDispatcher.platformBrightness;
     return themeMode == ThemeMode.system
         ? platformBrightness == Brightness.dark
         : themeMode == ThemeMode.dark;
@@ -105,8 +104,6 @@ class _PlatformThemeState extends State<PlatformTheme>
 
   @override
   Widget build(BuildContext context) {
-    print(
-        'PLATFORM THEME BUILD ${SchedulerBinding.instance.window.platformBrightness}');
     if (_matchCupertinoSystemChromeBrightness) {
       if (isCupertino(context)) {
         SystemChrome.setSystemUIOverlayStyle(